Flask_ML/requirements.txt
Flask
Flask_restful
matplotlib
Pillow
azure-storage-blob
tensorflow
numpy
matplotlib